Fix the Event Viewer to work  Asset Database (FastMode) build script
The Event Viewer used to work with the Asset Database build script, on the previous 1.8.5 verified version, and it broke at some point after that.
The FastModeInitializationOperation class is completely missing the code to initialize the event diagnostics if the postProfilerEvents option is enabled in the settings. This adds that code so the Event Viewer works again, using the same pattern as the InitializationOperation class.